+++ builds/cdn.js
@@ -1,0 +1,5 @@
+import clipboard from '../src/index'
+
+document.addEventListener('alpine:initializing', () => {
+    clipboard(window.Alpine)
+})
