+++ rollup.config.js
@@ -4,7 +4,7 @@ import filesize from 'rollup-plugin-filesize'
 import resolve from 'rollup-plugin-node-resolve'
 
 export default {
-    input: 'src/index.js',
+    input: 'builds/cdn.js',
     output: [
         {
             name: 'AlpineClipboard',
